% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_function.R
\name{ABC_rep}
\alias{ABC_rep}
\title{library(mvtnorm)
Function to run SMC ABC
Created by Carla Chen and Sarah Belet}
\usage{
ABC_rep(tol = 101, Np, alpha, x, obs, c_v)
}
\arguments{
\item{tol}{tolerance level}

\item{Np}{Number of particles}

\item{alpha}{proportion of particles dropped each step}

\item{x}{thingo vector}

\item{obs}{observations vector}

\item{c_v}{Probability that a particle is not moved at least once
---- Initial setup ---- #
Set initial tolerance
---- Rejection ABC part --------------}
}
\description{
library(mvtnorm)
Function to run SMC ABC
Created by Carla Chen and Sarah Belet
}
