% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev20.R
\name{bb_to_matchedIDs}
\alias{bb_to_matchedIDs}
\title{Determines the grid ID number for each grid space within a bounding box.
Useful for, eg, large areas of greenery that we can easily draw a boundary
   around, instead of grabbing the lat/long of every point within the region.
Returns a dataframe of grid ID numbers that will be associated with that land type.}
\usage{
bb_to_matchedIDs(bb, gridSize, grid)
}
\arguments{
\item{bb}{Bounding box of region, in lat/long}

\item{gridSize}{Grid size. Assumes square grid. In decimal form (eg 0.00025 = 25m^2 grid)}

\item{grid}{Dataframe of area grid, as per what's created in make_grid}
}
\value{
A dataframe of grid IDs associated with that land type.
}
\description{
Determines the grid ID number for each grid space within a bounding box.
Useful for, eg, large areas of greenery that we can easily draw a boundary
   around, instead of grabbing the lat/long of every point within the region.
Returns a dataframe of grid ID numbers that will be associated with that land type.
}
