% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{calculate_daily_EKS}
\alias{calculate_daily_EKS}
\title{New method of calculating Enzyme Kinetic Score for daily update (12-09-20)
This function takes in a vector of temperatures in Kelvin,
after a temperature deviate has been applied (if applicable).
Equation is from Focks 93.}
\usage{
calculate_daily_EKS(kelvinTemps, stage, const)
}
\arguments{
\item{kelvinTemps}{Vector of agent temperatures, in Kelvin}

\item{stage}{Development stage of agent, where:
\enumerate{
\item Egg
\item Larvae
\item Pupae
\item Adult
}}
}
\value{
A vector of daily enzyme kinetic scores for each agent.
}
\description{
New method of calculating Enzyme Kinetic Score for daily update (12-09-20)
This function takes in a vector of temperatures in Kelvin,
after a temperature deviate has been applied (if applicable).
Equation is from Focks 93.
}
