% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{find_trapped}
\alias{find_trapped}
\title{Finds any possible adult mosquitoes that are trapped by a particular trap
for a particular timestep.
If mosquitoes are within distance phi of a trap, there is a 0.5 probability
that a mosquito is trapped.
Only adults can be trapped.}
\usage{
find_trapped(traplat, traplong, phi, mozzie.dt, prob)
}
\arguments{
\item{traplat}{Lat of a particular trap.}

\item{traplong}{Long of a particular trap.}

\item{phi}{Distance that a mosquito is susceptible to a trap.}

\item{mozzie.dt}{Data.table of adult mosquitoes.}

\item{prob}{Probability of being trapped when within trap radius.}
}
\value{
Vector of mosquito IDs to be trapped. Expect to be length 0 for the first few days. If length 0 return -1.
}
\description{
Finds any possible adult mosquitoes that are trapped by a particular trap
for a particular timestep.
If mosquitoes are within distance phi of a trap, there is a 0.5 probability
that a mosquito is trapped.
Only adults can be trapped.
}
