% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev_functions.R
\name{get_gridID}
\alias{get_gridID}
\title{Given a latitude and longitude, determines the corresponding grid ID.
Assumes grid.df is a global variable.}
\usage{
get_gridID(testlat, testlong)
}
\arguments{
\item{testlat}{Latitude being considered.}

\item{testlong}{Longitude being considered.}
}
\value{
Corresponding grid ID number.
}
\description{
Given a latitude and longitude, determines the corresponding grid ID.
Assumes grid.df is a global variable.
}
