% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{grid_to_land_type}
\alias{grid_to_land_type}
\title{Grabs the land type of a mosquito given its grid ID.
Used for microclimates work.}
\usage{
grid_to_land_type(gridID)
}
\arguments{
\item{gridID}{Grid ID of an agent}
}
\value{
Land type
}
\description{
Grabs the land type of a mosquito given its grid ID.
Used for microclimates work.
}
