% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{init_juv_stage}
\alias{init_juv_stage}
\title{Roughly determines the stage of a juvenile clutch given numerical 'age'.
Needs checking with literature, or replace with better method.}
\usage{
init_juv_stage(age)
}
\arguments{
\item{age}{'Age' of clutch in days.}
}
\value{
Juvenile stage: 0 for egg, 1 for larval, 3 for pupal.
}
\description{
Roughly determines the stage of a juvenile clutch given numerical 'age'.
Needs checking with literature, or replace with better method.
}
\examples{
init_juv_stage(1)
init_juv_stage(7)
init_juv_stage(13)
}
