% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{init_position}
\alias{init_position}
\title{Generates random 'starting' positions for adult agents.
Uses boundary data for simulation region.
Draws a polygon from the boundary points given in boundaryDat
    and picks N randomly spaced points within it.
Includes a boundary check:
  will keep drawing points until there are N in the region/polygon.
Requires package 'sp' for Polygon object.}
\usage{
init_position(boundaryDat, N)
}
\arguments{
\item{boundaryDat}{Geographic boundary of simulation region.}

\item{N}{Number of agents.}
}
\value{
A data frame of lat and long positions for each adult agent.
}
\description{
Generates random 'starting' positions for adult agents.
Uses boundary data for simulation region.
Draws a polygon from the boundary points given in boundaryDat
    and picks N randomly spaced points within it.
Includes a boundary check:
  will keep drawing points until there are N in the region/polygon.
Requires package 'sp' for Polygon object.
}
