% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{initialise_adults}
\alias{initialise_adults}
\title{Creates an initial data.table of adult agents.
Since this is an initial spread, some data is made up/dummy:
for example, we don't know the ID of the parents.
Latitude/Longitude of agents is randomly determined through
another function \code{init_position} and added to the data.table
after construction.}
\usage{
initialise_adults(N, pmale)
}
\arguments{
\item{N}{The number of initial adult agents.}

\item{pmale}{Probability of being male.}
}
\value{
A data.table of \code{N} adult agents.
}
\description{
Creates an initial data.table of adult agents.
Since this is an initial spread, some data is made up/dummy:
for example, we don't know the ID of the parents.
Latitude/Longitude of agents is randomly determined through
another function \code{init_position} and added to the data.table
after construction.
}
\section{Data.table variables and initialisation}{

\itemize{
\item{ID:     Unique ID number of agent.}}
gender: Male is 0, female is 1. Sampled by 1 Binomial trial as opposed to a
        Bernoulli trial as Bernoulli requires another package.
mateID: Unique ID of their mate. Since no initial mosquitoes will have a mate
        yet, it is initialised as -1.
        Males will always have mateID as -1 since they can have multiple mates.
enzyme: Enzyme Kinetic Score. See \code{init} for explanation. Initialised
        uniform randomly.
age:    Age in days. Initialised uniform randomly around what we would expect
        young adults to be.
gonoCycle: Gonotrophic cycle. Means something different for males and females.
           males: number of times they've mated in a day, to be reset daily
           females: how many times they've laid a clutch of eggs
           we roughly estimate the gonoCycle of females based on \code{age}.
timeDeath: Timestep they died: initialised as -1 as they are alive.
typeDeath: Random mortality/trapped death/death due to old age: which type?
whereTrapped: in the event of trapped death, where did they die? -1 otherwise.
motherID: Unique ID of mother. -1 for initial adult data.table
fatherID: Unique ID of father. -1 for initial adult data.table.
infStatus: 1 if they carry Wolbachia, 0 if no Wolbachia, -1 for CI
           for initial wild type we assume they all start with 0.
lat:      Initial north/south or 'y' coordinate of agent. Should start with -16.
long:     Initial east/west of 'x' coordinate of agent. Should start with 145.
}

