% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{initialise_eggs}
\alias{initialise_eggs}
\title{Creates a data.table of newly laid eggs to add to the juvenile data.table.
Input is the indices of mothers attempting to lay eggs.
Much of this code is regarding the handling of CI:
\begin{itemize}{
 \item{Neither mother nor father carry Wolbachia: no Wolbachia in offspring}
 \item{Mother carries Wolbachia: offspring have nonzero probability of carrying}
 \item{Mother does not carry, father does: offspring suffer CI and won't hatch}
}
Latitude/Longitude of agents is determined as the same position that the
 mother is currently at.
Clutch sizes, as per literature, differ depending on Wolbachia status of mother.
This is handled by the ABC parameter eta_1 and eta_2.}
\usage{
initialise_eggs(toLay)
}
\arguments{
\item{toLay}{The indices of mothers attempting to lay a clutch.}
}
\value{
A data.table of juvenile agents in stage 1 corresponding to each mother in \code{toLay}.
}
\description{
Creates a data.table of newly laid eggs to add to the juvenile data.table.
Input is the indices of mothers attempting to lay eggs.
Much of this code is regarding the handling of CI:
\begin{itemize}{
 \item{Neither mother nor father carry Wolbachia: no Wolbachia in offspring}
 \item{Mother carries Wolbachia: offspring have nonzero probability of carrying}
 \item{Mother does not carry, father does: offspring suffer CI and won't hatch}
}
Latitude/Longitude of agents is determined as the same position that the
 mother is currently at.
Clutch sizes, as per literature, differ depending on Wolbachia status of mother.
This is handled by the ABC parameter eta_1 and eta_2.
}
\section{Data.table variables and initialisation}{

motherID:   ID of mother.
fatherID:   ID of father.
age:        Age in days. Initialised at 0 since they're new!
stage:      Development stage of clutch. Should only be 1 since these are eggs.
infProb:    Probability of carrying Wolbachia. 0: no Wolbachia, -1: Cytoplasmic
              Incompatability, else infProb is nonzero.
lat:        Initial north/south or 'y' coordinate of agent.
              Should start with -16. Should be same as mother.
long:       Initial east/west of 'x' coordinate of agent.
              Should start with 145. Should be same as mother.
clutchSize: Number of juveniles in the clutch.
pDeath:     Probability of death.
}

