% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev20.R
\name{initialise_enzyme_wmicroclim}
\alias{initialise_enzyme_wmicroclim}
\title{Constructs the EKM chart with microclimates.
Uses data about the area of the region we're simulating.
If you're using this function for a different area, you will need to change
 all of the data that's read in in the first part of the function.
The lookup table is deterministic, and is made of nested lists.
The order will be: [stage][land type][day]
So when we update the EKS for each agent, we add the value that is at df[stage][land type][day]
Eg: for an egg that is in land type 1 at time step t=2, we add the value at dt[0][1][2] to the agents' EKS
To use this list, indexing works like: EKM_chart[[stage]][land_type, ][day]
Land types (as of 16/1/19) are:
\enumerate{
\item{Uncovered greenery}
\item{Vegetation}
\item{Roads}
\item{Water bodies}
\item{Carpark/uncovered concrete}
\item{Large buildings}
\item{Houses (tempdev informed by house type)}
}}
\usage{
initialise_enzyme_wmicroclim(bdary, tempData, typeTempDevs, noDays,
  noLandTypes, gridSize, const)
}
\arguments{
\item{bdary}{Boundary of simulation region.}

\item{tempData}{Vector of average temperatures over the simulation period.}

\item{typeTempDevs}{Temperature deviates for each land type.}

\item{noDays}{Number of days in the simulation period.}

\item{noLandTypes}{Number of different land types we consider.}

\item{gridSize}{Grid size, as per make_grid and bb_to_matchedIDs functions}

\item{const}{Constants for EKM calculation.}
}
\value{
A list of 4 dataframes for each development stage, of the EKS for each day of the simulation per each land type
}
\description{
Constructs the EKM chart with microclimates.
Uses data about the area of the region we're simulating.
If you're using this function for a different area, you will need to change
 all of the data that's read in in the first part of the function.
The lookup table is deterministic, and is made of nested lists.
The order will be: [stage][land type][day]
So when we update the EKS for each agent, we add the value that is at df[stage][land type][day]
Eg: for an egg that is in land type 1 at time step t=2, we add the value at dt[0][1][2] to the agents' EKS
To use this list, indexing works like: EKM_chart[[stage]][land_type, ][day]
Land types (as of 16/1/19) are:
\enumerate{
\item{Uncovered greenery}
\item{Vegetation}
\item{Roads}
\item{Water bodies}
\item{Carpark/uncovered concrete}
\item{Large buildings}
\item{Houses (tempdev informed by house type)}
}
}
