% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{initialise_graveyard}
\alias{initialise_graveyard}
\title{Initialises graveyard data.table
The graveyard is where the entries of all dead adult mozzies go}
\usage{
initialise_graveyard()
}
\description{
Initialises graveyard data.table
The graveyard is where the entries of all dead adult mozzies go
}
