% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{initialise_juv_graveyard}
\alias{initialise_juv_graveyard}
\title{Initialises juvenile graveyard data.table
The entries here are for egg clutches that won't hatch due to CI
This does not include juveniles that died of natural causes
  (since that's just a number in the juv.dt data.table)
 @return Initialised data.table of the juvenile graveyard.}
\usage{
initialise_juv_graveyard()
}
\description{
Initialises juvenile graveyard data.table
The entries here are for egg clutches that won't hatch due to CI
This does not include juveniles that died of natural causes
  (since that's just a number in the juv.dt data.table)
 @return Initialised data.table of the juvenile graveyard.
}
