% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{initialise_juveniles}
\alias{initialise_juveniles}
\title{Creates an initial data.table of juvenile agents.
Since this is an initial spread, some data is made up/dummy:
for example, we don't know the ID of the parents.
Latitude/Longitude of agents is randomly determined and added to the
 data.table after construction.}
\usage{
initialise_juveniles(Njuv, param, grid.df, boundaryDat)
}
\arguments{
\item{Njuv}{The number of initial juvenile agents.}
}
\value{
A data.table of Njuv juvenile agents.
}
\description{
Creates an initial data.table of juvenile agents.
Since this is an initial spread, some data is made up/dummy:
for example, we don't know the ID of the parents.
Latitude/Longitude of agents is randomly determined and added to the
 data.table after construction.
}
\section{Data.table variables and initialisation}{

\describe{
\item{motherID:}{ID of mother.}
\item{fatherID:}{ID of father.}
\item{age:}{Age in days. Initialised uniform randomly around what we would expect
             young adults to be.}
\item{stage:}{Development stage of clutch. 1: egg, 2: larvae, 3: pupae.}
\item{infProb:}{Probability of carrying Wolbachia. 0: no Wolbachia, -1: Cytoplasmic
              Incompatability, else infProb is nonzero.}
\item{lat:}{Initial north/south or 'y' coordinate of agent. Should start with -16.}
\item{long:}{Initial east/west of 'x' coordinate of agent. Should start with 145.}
\item{clutchSize:}{Number of juveniles in the clutch.}
\item{enzyme:}{Enzyme Kinetic Score of agents. Initialised uniform randomly. FIX}
\item{pDeath:}{Probability of death.}
}
}

