% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{initialize_enzyme}
\alias{initialize_enzyme}
\title{Initialises the Enzyme Kinetic model lookup table for the model run.
The EKM is how we dictate the development rate of agents.
The EKM model is the same one used in Focks 1993.
Input data is in CELSIUS, EKM works with KELVIN.}
\usage{
initialize_enzyme(dailyTemps)
}
\arguments{
\item{dailyTemps}{Dataframe of daily average temperatures as output from
function temperature_chart}
}
\value{
Chart of EKM updates for each development stage for each day.
}
\description{
The model uses different sets of experimentally-derived parameters
for each four development stage: Egg, Larvae, Pupae, Gonotrophic.
Hence there are four equations.
}
\section{Notation and constants}{

KELV_CONV: Conversion constant between Kelvin <-> Celcius.
R: Universal gas constant.
All other constants are experimentally-determined constants
which can be found in Focks 1993.
Suffixes E, L, P, G for the constants refer to each of the
development stages.
}

\section{"Why do you multiply by 24?"}{

We multiply by 24 since timesteps are in days, and EKM model
is in timestep unit of hours.
}

