% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_steps.R
\name{juv_emergence}
\alias{juv_emergence}
\title{Runs the 'juvenile emergence' step.
Determines which juvenile agents are to age into adult agents,
  and calls \code{juv_to_adult} on them.
IMPORTANT: this should update \code{mozzie.dt} and \code{juv.dt} on a global scale.}
\usage{
juv_emergence(juv.dt)
}
\arguments{
\item{juv.dt}{The juvenile data.table.}
}
\description{
Runs the 'juvenile emergence' step.
Determines which juvenile agents are to age into adult agents,
  and calls \code{juv_to_adult} on them.
IMPORTANT: this should update \code{mozzie.dt} and \code{juv.dt} on a global scale.
}
