% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_steps.R
\name{juv_stage_update}
\alias{juv_stage_update}
\title{Determines the juvenile agents that develop into the next stage, and
  updates the development stage of these juveniles.
Once \code{enzyme} exceeds 0.95, a juvenile agent develops to the next stage.
Recall that Stage 1: egg, 2: larvae, 3: pupae.
Any pupae that are to develop into the next stage, i.e. adulthood, should
  have been handled by \code{juv_emergence}.
IMPORTANT: this should update \code{juv.dt} on a global scale.}
\usage{
juv_stage_update(juv.dt)
}
\arguments{
\item{juv.dt}{The juvenile data.table.}
}
\description{
Determines the juvenile agents that develop into the next stage, and
  updates the development stage of these juveniles.
Once \code{enzyme} exceeds 0.95, a juvenile agent develops to the next stage.
Recall that Stage 1: egg, 2: larvae, 3: pupae.
Any pupae that are to develop into the next stage, i.e. adulthood, should
  have been handled by \code{juv_emergence}.
IMPORTANT: this should update \code{juv.dt} on a global scale.
}
