% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{juv_to_adult}
\alias{juv_to_adult}
\title{Takes a juvenile agent and splits it into adult ages
This is to simulate 'emergence' of juveniles form aquatic stage.
Many variables stay the same, like motherID, but new ones are added or changed
Agents also undergo one motility event when they emerge.
This becomes their new lat/long.}
\usage{
juv_to_adult(juvID, idStart, pmale, lambda)
}
\arguments{
\item{juvID}{The index of the agent in the juvenile data.table to be converted to adult agents.}

\item{idStart}{Where to start 'counting' the ID numbers of new agents from.}

\item{pmale}{Probability of being male.}

\item{lambda}{Shape parameter for distance calculation. ABC parameter.}
}
\value{
A data.table of \code{N} adult agents.
}
\description{
Takes a juvenile agent and splits it into adult ages
This is to simulate 'emergence' of juveniles form aquatic stage.
Many variables stay the same, like motherID, but new ones are added or changed
Agents also undergo one motility event when they emerge.
This becomes their new lat/long.
}
\section{Data.table variables and initialisation}{

ID:     Unique ID number of agent.
gender: Male is 0, female is 1. Sampled by 1 Binomial trial as opposed to a
        Bernoulli trial as Bernoulli requires another package.
mateID: Unique ID of their mate. Since no new adults will have a mate
        yet, it is initialised as -1.
        Males will always have mateID as -1 since they can have multiple mates.
enzyme: Enzyme Kinetic Score. See \code{init} for explanation. Initialised
        at 0 since they have just moved up from previous stage.
age:    Age in days. Given from their juvenile agent entry.
gonoCycle: Gonotrophic cycle. Means something different for males and females.
           males: number of times they've mated in a day, to be reset daily
           females: how many times they've laid a clutch of eggs
           Starts at 0 for everyone since they've just emerged.
timeDeath: Timestep they died: initialised as -1 as they are alive.
typeDeath: Random mortality/trapped death/death due to old age: which type?
whereTrapped: in the event of trapped death, where did they die? -1 otherwise.
motherID: Unique ID of mother. Taken from their juvenile agent entry.
fatherID: Unique ID of father. Taken from their juvenile agent entry.
infStatus: 1 if they carry Wolbachia, 0 if no Wolbachia, -1 for CI
           There should not be any with -1 because they should not have hatched.
lat:      Initial north/south or 'y' coordinate of agent.
           Should start with 145. Determined by motility event.
long:     Initial east/west of 'x' coordinate of agent.
           Should start with -16. Determined by motility event.
}

