% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev20.R
\name{make_grid}
\alias{make_grid}
\title{Converts a polygon boundary into a grid mesh, but in dataframe form}
\usage{
make_grid(bdary, gridSize)
}
\arguments{
\item{bdary}{List of points along the boundary of the simulation area (in lat/long)}

\item{gridSize}{Grid size. Assumes square grid. In decimal form (eg 0.00025 = 25m^2 grid)}
}
\value{
A dataframe of lat/long coordinates representing the region described by bdary
        with space sizes gridSize.
}
\description{
Converts a polygon boundary into a grid mesh, but in dataframe form
}
