% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{make_plot_data}
\alias{make_plot_data}
\title{Makes plot data for one graveyard dataframe.
Because I'm tired of copying and pasting the same code over and over...}
\usage{
make_plot_data(grav.data)
}
\arguments{
\item{grav.data}{A graveyard data.frame.}
}
\description{
Makes plot data for one graveyard dataframe.
Because I'm tired of copying and pasting the same code over and over...
}
