% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run_func.R
\name{model_run_new}
\alias{model_run_new}
\title{God i don't know hopefully this will work.
This version does NOT run microclimates.
Input is the ABC parameters
Output are vectors of interest over time.}
\usage{
model_run_new(lambda, k, phi, alpha_a, alpha_j, eta)
}
\arguments{
\item{lambda}{Shape param for dispersal}

\item{k}{Parameter for mating}

\item{phi}{Parameter for trapping}

\item{alpha_a}{Mortality rate of adult mosquitoes}

\item{alpha_j}{Mortality rate of juvenile mosquitoes}

\item{eta}{Clutch size of each egg-laying event.
Option to select either high, low, or medium fixed values for each parameter
comment this all out after the plots are done.
low = -1, medium = 0, high = 1.
Mid parameters:
End mid parameters
Low parameters:
End low parameters
High parameters:
End high parameters
This section sets up alternative temperature deviate setup methods ----
There are three alternative TD setups for testing purposes:
1) TDs assigned uniformly at random
2) TDs assigned proportionally randomly to how frequently they appear
   in the "realistic" microclimates setup
3) The grid is partitioned into the different land types, where the size
   of each partition is proportional to how frequently they appear in the
   "realistic" microclimates setup. (The order that the partitions appear
   in is randomly generated).
There are three function that do the above. The "tempdev_type" variable can
be changed to 0 (keep realistic), 1 (option 1), 2 (option 2) and 3 (option 3).
The functions assign land type, then we match it to the associated temperature
deviate below.
End alternative temperature deviate methods ----
---- Set WD for saving graveyard ----
Each row will be a timestep, each column will be the time each step takes
PLUS the number of agents.
------------------------

ADDED 12/11/20: EGG HATCH RATE OF 0.8 AS PER PERRAN ROSS

This won't error if length(to.adult) == 0 hopefully
------------------------

22-09-20: reduce proportion of females laying eggs per day: CHECK
ADDED 14/11/20: MAKING JUV DEATH RATE HIGHER FOR THOSE WITH WB
add to graveyard
remove from mozzie dataframe
TESTING 05/11/20: RELEASED ADULTS WITH HIGHER DEATH RATE HHHHHHH --------
END TESTING 05/11/20 HHHH ----------------------
Begin Step 6 Take 2: a faster way to update microclimates ----
For adults/mozzie.dt: everyone has the same EKS update!
For juveniles, it's a little bit more involved:
Get the temperature deviate for each mozzie
Apply the temperature deviate to today's temperature and then convert to Kelvin
This is the temperature that will then be fed into the EKM
Find EKS score
first: for adults, stage = 4
Find the EKS for each adult for today
Add to the current enzyme
now: for juveniles
DOBULE CHECK the below works....
commented the below out 7/1/20



End Step 6 Take 2 :)
dataframe for optimisation:
MATING ATTEMPT NUMBER 3...... I S2G THIS HAS TO BE THE LAST TIME
we do a vectorised for loop. the find_a_mate function works on one agent at a time.
Get a list of females who can mate.
There's a ~30% chance that those females who can mate, will mate.
apply function to find mates
Update female's mate ID
Update male's gonoCycle
Set up stuff to get graveyard data
end graveyard data
capture data
save graveyard as a csv}
}
\description{
God i don't know hopefully this will work.
This version does NOT run microclimates.
Input is the ABC parameters
Output are vectors of interest over time.
}
