% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{mosquito_trapping}
\alias{mosquito_trapping}
\title{Trapping of mosquitoes.}
\usage{
mosquito_trapping(trapID, phi, mozzie.dt)
}
\arguments{
\item{trapID}{unique ID of trap}

\item{phi}{phi/2 is the radius of influence of the trap.}

\item{mozzie.dt}{data.table of adult agents}
}
\value{
List of mosquito IDs that are to be trapped.
}
\description{
This algorithm cycles through every trap in the region that is
active at time t. Every adult mosquito within 'radius' phi/2 of the
trap has a 50% chance of being trapped.
('Radius' here referring to a square of lenth phi.)
Returns list of mosquito IDs that are to be trapped.
}
