% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev20.R
\name{partition_tempdev}
\alias{partition_tempdev}
\title{Alternative temperature deviate scheme 3:
The grid is partitioned into land types, proportional in size
to the land type's frequency in the "realistic" land types.
The order in which each partition appears is randomly generated.
This requires initialise_enzyme_wmicroclim to have run first, and
landtype.df must be loaded into the global environment.}
\usage{
partition_tempdev(landtype.df, noLandTypes, typeTempDevs)
}
\arguments{
\item{landtype.df}{Dataframe of each grid space lat/long, their "realistic" land type,
their "realistic" temperature deviate, and their grid ID.}

\item{noLandTypes}{Number of land types.}

\item{typeTempDevs}{Vector of temperature deviates. Must be of length noLandTypes.}
}
\value{
A vector of randomly assigned LandTypes.
}
\description{
Alternative temperature deviate scheme 3:
The grid is partitioned into land types, proportional in size
to the land type's frequency in the "realistic" land types.
The order in which each partition appears is randomly generated.
This requires initialise_enzyme_wmicroclim to have run first, and
landtype.df must be loaded into the global environment.
}
