% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_positions}
\alias{plot_positions}
\title{Plots positions of mosquitoes in simulation region.}
\usage{
plot_positions(mozzie.dt)
}
\arguments{
\item{mozzie.dt}{data.table of adult agents}
}
\value{
A plot of mozzie positions.
}
\description{
Due to the number of agents in the simulation, this is not
a good summary plot. Try a density plot instead.
}
