% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_proportions_by_buf}
\alias{plot_proportions_by_buf}
\title{Plots proportion of mosquitoes trapped that have Wolbachia over time for a buffer zone.}
\usage{
plot_proportions_by_buf(graveyardTrapped, buf)
}
\arguments{
\item{graveyardTrapped}{Subset of graveyard for trapped mosquitoes only. This is modified graveyard
data with only columns whereTrapped, infStatus, timeDeath, Buf and count}

\item{buf}{Buffer zone to plot.}
}
\value{
A plot of trapped Wolbachia proportions over time for a given buffer zone.
}
\description{
The main buffers we should care about is 0-100m, 100-200m, then 200m and beyond.
}
