% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev20.R
\name{points_to_matchedIDs}
\alias{points_to_matchedIDs}
\title{Determines the grid ID number corresponding to each lat/long point in the input.
Returns a dataframe of grid ID numbers that will be associated with that land type.}
\usage{
points_to_matchedIDs(points, grid)
}
\arguments{
\item{points}{Lat/long points to be matched to a grid ID.}

\item{grid}{Dataframe of area grid, as per what's created in make_grid}
}
\value{
A dataframe of grid IDs associated with that land type.
}
\description{
Determines the grid ID number corresponding to each lat/long point in the input.
Returns a dataframe of grid ID numbers that will be associated with that land type.
}
