% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_steps.R
\name{produce_eggs}
\alias{produce_eggs}
\title{Determines which mothers are ready to lay eggs and calls \code{initialise_eggs}.
Conditions to be satisfied for a mother to lay a clutch of eggs:
\enumerate{
 \item {Has a mate}
 \item {Enzyme Kinetic Score > 1 and gonoCycle = 0 OR}
 \item {EKS > 1.58 and gonoCycle = 1 OR}
 \item {EKS > 2.16 and gonoCycle = 2.}
}}
\usage{
produce_eggs(mozzie.dt)
}
\arguments{
\item{mozzie.dt}{data.table of adult agents.}
}
\description{
IMPORTANT: This should update \code{juv.dt} on a global scale.
}
