% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev20.R
\name{proportional_random_tempdev}
\alias{proportional_random_tempdev}
\title{Alternative temperature deviate scheme 2:
Temperature deviates picked randomly,
but proportional to their prevalence in the "realistic" tempdev scheme/
This requires initialise_enzyme_wmicroclim to have run first, and
landtype.df must be loaded into the global environment.}
\usage{
proportional_random_tempdev(landtype.df, noLandTypes, typeTempDevs)
}
\arguments{
\item{landtype.df}{Dataframe of each grid space lat/long, their "realistic" land type,
their "realistic" temperature deviate, and their grid ID.}

\item{noLandTypes}{Number of land types.}

\item{typeTempDevs}{Vector of temperature deviates. Must be of length noLandTypes.}
}
\value{
A vector of randomly assigned LandTypes.
}
\description{
Alternative temperature deviate scheme 2:
Temperature deviates picked randomly,
but proportional to their prevalence in the "realistic" tempdev scheme/
This requires initialise_enzyme_wmicroclim to have run first, and
landtype.df must be loaded into the global environment.
}
