% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{release_clean}
\alias{release_clean}
\title{Clean release data.
Since the method for estimating mosquito numbers is inprecise,
there are non-integer numbers of mosquitoes in each column.
Furthermore, this data set has no dates, rather the "day" as per Carla's work,
so I'm adding the dates back in.}
\usage{
release_clean(releaseDat)
}
\arguments{
\item{releaseDat}{Mosquito release data.}
}
\value{
Dataframe of cleaned release data.
}
\description{
Clean release data.
Since the method for estimating mosquito numbers is inprecise,
there are non-integer numbers of mosquitoes in each column.
Furthermore, this data set has no dates, rather the "day" as per Carla's work,
so I'm adding the dates back in.
}
