% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{resize_clutch}
\alias{resize_clutch}
\title{Applies daily natural death rate to a juvenile clutch.}
\usage{
resize_clutch(clutch.size, alpha)
}
\arguments{
\item{clutch.size}{Size of clutch.}

\item{alpha}{Natural death rate.}
}
\value{
Updated \code{clutch.size}.
}
\description{
Uses the exponential model of mortality (constant death rate at each timestep)
}
