% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{temperature_chart}
\alias{temperature_chart}
\title{Calculates the average daily temperature over the simulation period.
Temperature calculations are in Celcius.}
\usage{
temperature_chart(mindf, maxdf, noTimeSteps)
}
\arguments{
\item{mindf}{Raw data of all minimum daily temperatures, read into a dataframe.}

\item{maxdf}{Raw data of all maximum daily temperatures, read into a dataframe.}

\item{noTimeSteps}{Number of timesteps in the simulation.}
}
\value{
The mean temperature for each day of the simulation period, expressed as
a dataframe of length \code{noTimeSteps}.
}
\description{
Function splits the data into separate months, lays it end to end,
and then calculates the mean temperature for each day of the
simulation period.
}
\details{
Note that due to months not having the same number of days, we remove
any -1 entries which correspond to days like Feb 30.

This function is deterministic and should yield the same result for
the same input.
}
