% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{trap_clean}
\alias{trap_clean}
\title{Clean trapping data.
Adds 3 columns at the end of the data CURRENTLY UNIMPLEMENTED:
\describe{
 \item{FWProp}{Proportion of trapped females that carry Wolbachia}
 \item{MWProp}{Proportion of trapped males that carry Wolbachia}
 \item{WProp}{Proportion of trapped mosquitoes that carry Wolbachia}
}
Adapted from code I wrote to create plots.}
\usage{
trap_clean(trapped)
}
\arguments{
\item{trapped}{Input trapping data, as a dataframe}
}
\value{
Dataframe of cleaned trapping data.
}
\description{
Clean trapping data.
Adds 3 columns at the end of the data CURRENTLY UNIMPLEMENTED:
\describe{
 \item{FWProp}{Proportion of trapped females that carry Wolbachia}
 \item{MWProp}{Proportion of trapped males that carry Wolbachia}
 \item{WProp}{Proportion of trapped mosquitoes that carry Wolbachia}
}
Adapted from code I wrote to create plots.
}
