% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{trap_data_setup}
\alias{trap_data_setup}
\title{Set up output file for trapping
It's going to be in longform
Variables: Lat, Long, Sex, Day, Wolbachia status, Number trapped.}
\usage{
trap_data_setup(trapped, traploc, number_of_traps, noTimeSteps)
}
\arguments{
\item{trapped}{Cleaned trapping data}

\item{traploc}{Vector of trap locations}

\item{number_of_traps}{Number of traps.}

\item{noTimeSteps}{Number of time steps.}
}
\value{
Longform dataframe to record trapping data.
}
\description{
Set up output file for trapping
It's going to be in longform
Variables: Lat, Long, Sex, Day, Wolbachia status, Number trapped.
}
