% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{trap_empty}
\alias{trap_empty}
\title{Gets Wolbachia infection proportion of traps at appropriate time intervals.
In reality, traps were delpoyed and then emptied on certain days.
So we can't just look at the proportion of Wolbachia-carriers over time for the
entire graveyard}
\usage{
trap_empty(trapClean)
}
\arguments{
\item{trapClean}{Cleaned trapping data as per function trap_clean.}
}
\description{
Gets Wolbachia infection proportion of traps at appropriate time intervals.
In reality, traps were delpoyed and then emptied on certain days.
So we can't just look at the proportion of Wolbachia-carriers over time for the
entire graveyard
}
