% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{trap_setup}
\alias{trap_setup}
\title{Sets up traps:
Determines where unique traps positions are.}
\usage{
trap_setup(trapped)
}
\arguments{
\item{trapped}{Dataframe of trap data as per function trap_clean}
}
\value{
Dataframe of trap locations in lat/long.
}
\description{
Sets up traps:
Determines where unique traps positions are.
}
