% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_enzyme}
\alias{update_enzyme}
\title{Gets the amount by which the Enzyme Kinetic Score for an agent will increase.
This value is added to their current EKS.
This function works as a 'lookup' for the EKS data, which is all calculated
    at the beginning of the simulation- this function just grabs it}
\usage{
update_enzyme(type, time)
}
\arguments{
\item{stage}{Stage of agent. 1: egg, 2: larvae, 3: pupae, 4: adult.}

\item{timestep}{Current timestep.}
}
\value{
Amount by which an agent's EKS will increase for that timestep.
}
\description{
Gets the amount by which the Enzyme Kinetic Score for an agent will increase.
This value is added to their current EKS.
This function works as a 'lookup' for the EKS data, which is all calculated
    at the beginning of the simulation- this function just grabs it
}
\examples{
update_enzyme(1, 1)
update_enzyme(4, 13)
}
