% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_enzyme_microclim}
\alias{update_enzyme_microclim}
\title{Gets the amount by which the Enzyme Kinetic Score for an agent will increase.
This value is added to their current EKS.
This fuction is used when microclimates are enabled.
This function works as a 'lookup' for the EKS data, which is all calculated
    at the beginning of the simulation- this function just grabs it}
\usage{
update_enzyme_microclim(stage, timestep, landType)
}
\arguments{
\item{stage}{Stage of agent. 1: egg, 2: larvae, 3: pupae, 4: adult.}

\item{timestep}{Current timestep.}

\item{landType}{Land Lype that the agent is on.}

\item{grid.df}{Lookup table that includes the land type}
}
\value{
Amount by which an agent's EKS will increase for that timestep.
}
\description{
Gets the amount by which the Enzyme Kinetic Score for an agent will increase.
This value is added to their current EKS.
This fuction is used when microclimates are enabled.
This function works as a 'lookup' for the EKS data, which is all calculated
    at the beginning of the simulation- this function just grabs it
}
