% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_juv_stage}
\alias{update_juv_stage}
\title{Updates the development stage of juvenile agents.}
\usage{
update_juv_stage(stage, enzyme)
}
\arguments{
\item{stage}{List of all juvenile stages from juv.dt.}

\item{enzyme}{List of all juvenile agent enzyme score from juv.dt.}
}
\value{
Updated \code{stage} and \code{enzyme} of agents, as a list.
}
\description{
If a juvenile agent's Enzyme Kinetic Score is greater than 0.95,
they age up to the next development stage:
\itemize{
\item{Egg to larvae OR}
\item{Larvae to pupae.}
}
Pupae development into adult agents should have already been handled.
This function runs on the entire list of juvenile agents at once.
}
