% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{when_trap_active}
\alias{when_trap_active}
\title{Not all traps are active during the entire simulation.
This function determines if a trap is active at time t before it attempts to trap anything.}
\usage{
when_trap_active(currentTraps)
}
\arguments{
\item{currentTraps}{Cleaned trapping data with active dates converted to numerical days}
}
\value{
Boolean TRUE if active and FALSE if not active.
}
\description{
Not all traps are active during the entire simulation.
This function determines if a trap is active at time t before it attempts to trap anything.
}
