% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_functions.R
\name{when_trap_empty}
\alias{when_trap_empty}
\title{Returns a vector of the dates where we're "cleaning" out the traps
i.e. when we actually "observe" the data in the graveyard}
\usage{
when_trap_empty(trapClean)
}
\arguments{
\item{trapClean}{Cleaned trapping data as per function trap_clean.}
}
\value{
Vector of days where we empty the traps.
}
\description{
Returns a vector of the dates where we're "cleaning" out the traps
i.e. when we actually "observe" the data in the graveyard
}
