% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{find_mate}
\alias{find_mate}
\title{Finds a mate for a female mosquitoes of breeding age.
We draw a box of 'radius' \code{k} around the agent.
\code{k} is an ABC parameter and usually between 11m to 100m.
Note that here we mainly deal with ID number, an actual data.table variable,
 and not the agent's index in the main data.table.}
\usage{
find_mate(femID, k, max_daily_mates)
}
\arguments{
\item{femID}{ID of female looking for mate.}

\item{k}{Distance within which female mosquito 'searches' for mate. ABC parameter.}

\item{max_daily_mates}{Number of times a male can mate in a day.}
}
\value{
\code{mateID} of agent, or -1 if agent does not find a mate.
}
\description{
Finds a mate for a female mosquitoes of breeding age.
We draw a box of 'radius' \code{k} around the agent.
\code{k} is an ABC parameter and usually between 11m to 100m.
Note that here we mainly deal with ID number, an actual data.table variable,
 and not the agent's index in the main data.table.
}
\examples{
add(1, 1)
add(10, 1)
}
