% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdev_functions.R
\name{get_gridID}
\alias{get_gridID}
\title{Given a latitude and longitude, determines the corresponding grid ID.
Assumes grid.df is a global variable.
This is to find the grid ID of an agent. assign_grid() is for microclimates.}
\usage{
get_gridID(testlat, testlong, gridlong, gridlat)
}
\arguments{
\item{testlat}{Latitude being considered.}

\item{testlong}{Longitude being considered.}
}
\value{
Corresponding grid ID number.
edited 22-11-20, grid.df gets read explicitly into get_gridID
}
\description{
Given a latitude and longitude, determines the corresponding grid ID.
Assumes grid.df is a global variable.
This is to find the grid ID of an agent. assign_grid() is for microclimates.
}
