% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{initialise_release}
\alias{initialise_release}
\title{Simulates a release of Wolbachia-carrying mosquitoes.
Releases in this experiment are always adults.
Much of this is functionally similar to initialise_adults.
Since this is a release, some data is made up/dummy:
for example, we don't know the ID of the parents.
Latitude/Longitude of agents is given by data supplied by the WMP.}
\usage{
initialise_release(noReleased, pmale, lat, long)
}
\arguments{
\item{noReleased}{The number of initial adult agents.}

\item{pmale}{Probability of being male.}

\item{lat}{Latitude of release site.}

\item{long}{Longitude of release site.}
}
\value{
A data.table of \code{N} adult agents.
}
\description{
Simulates a release of Wolbachia-carrying mosquitoes.
Releases in this experiment are always adults.
Much of this is functionally similar to initialise_adults.
Since this is a release, some data is made up/dummy:
for example, we don't know the ID of the parents.
Latitude/Longitude of agents is given by data supplied by the WMP.
}
\section{Data.table variables and initialisation}{

ID:     Unique ID number of agent.
gender: Male is 0, female is 1. We know proportions of female/males released
         from data.
mateID: Unique ID of their mate. Since no initial mosquitoes will have a mate
        yet, it is initialised as -1.
        Males will always have mateID as -1 since they can have multiple mates.
enzyme: Enzyme Kinetic Score. See \code{init} for explanation. Initialised
        uniform randomly.
age:    Age in days. Initialised uniform randomly around what we would expect
        young adults to be.
gonoCycle: Gonotrophic cycle. Means something different for males and females.
           males: number of times they've mated in a day, to be reset daily
           females: how many times they've laid a clutch of eggs
           we roughly estimate the gonoCycle of females based on \code{age}.
timeDeath: Timestep they died: initialised as -1 as they are alive.
typeDeath: Random mortality/trapped death/death due to old age: which type?
whereTrapped: in the event of trapped death, where did they die? -1 otherwise.
motherID: Unique ID of mother. -1 since wild release.
fatherID: Unique ID of father. -1 since wild release.
infStatus: 1 if they carry Wolbachia, 0 if no Wolbachia, -1 for CI.
           They should mostly be 1 with a few 0 due to incomplete transmission.
lat:      Initial north/south or 'y' coordinate of agent. Should start with -16.
long:     Initial east/west of 'x' coordinate of agent. Should start with 145.
}

