% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{random_dispersal}
\alias{random_dispersal}
\title{Random dispersal of a single agent according to distance formulae.}
\usage{
random_dispersal(lat, long, lambda)
}
\arguments{
\item{lat}{Latitude or 'y' coordinate of agent.}

\item{long}{Longitude or 'x' coordinate of agent.}

\item{lambda}{Shape parameter for distance calculation. ABC parameter.}
}
\value{
Updated \code{lat} and \code{long} of agent, as a list.
below uses DEGREES in angle for calculation and not radians
do we need to do anything about that???
}
\description{
Method we use: distance and bearing, Haversine method
Note we work in units of radians and kilometres.
\enumerate{
 \item{Calculate 'd', the distance travelled}
 \item{Calculate 'theta', the direction or bearing travelled}
 \item{Convert to lat and long.}
}
}
